p6 updates

+++ master/all.csh
@@ -1,7 +1,7 @@
 #!/bin/csh -f
-#SBATCH -J p31_11
-#SBATCH -e p31_11.err
-#SBATCH -o p31_11.out
+#SBATCH -J p6_11
+#SBATCH -e p6_11.err
+#SBATCH -o p6_11.out
 #SBATCH -t 7:55:00
 #SBATCH -q batch
 #SBATCH -A marine-cpu
@@ -12,12 +12,12 @@
 source /etc/profile.d/modules.csh
 setenv PATH /scratch2/NCEPDEV/climate/Robert.Grumbine/anaconda3/bin:$PATH
 
-setenv expt bm31
-setenv FCST_BASE /scratch2/NCEPDEV/climate/Lydia.B.Stefanova/Models/ufs_b31/SeaIce/ 
-setenv EXDIR /scratch2/NCEPDEV/climate/Robert.Grumbine/${expt}.verf
-cd /scratch2/NCEPDEV/climate/Robert.Grumbine/${expt}.verf
+setenv expt p6.0
+setenv FCST_BASE /scratch2/NCEPDEV/climate/Lydia.B.Stefanova/Models/ufs_p6/SeaIce/ 
+setenv EXDIR   /scratch2/NCEPDEV/climate/Robert.Grumbine/prototype_evaluations/${expt}.verf
+setenv RUNBASE /scratch2/NCEPDEV/stmp1/Robert.Grumbine/prototype_evaluations/${expt}.verf
+cd /scratch2/NCEPDEV/climate/Robert.Grumbine/prototype_evaluations/${expt}.verf
 setenv base `pwd`
-setenv RUNBASE /scratch2/NCEPDEV/stmp1/Robert.Grumbine/${expt}.verf
 
 echo env $FCST_BASE $EXDIR $base $RUNBASE
 
@@ -27,6 +27,7 @@ if ( ! -d $XDG_RUNTIME_DIR ) then
   mkdir -p -m 700 $XDG_RUNTIME_DIR
 endif
 echo $XDG_RUNTIME_DIR for python graphic support
+setenv MPLCONFIGDIR /scratch2/NCEPDEV/climate/Robert.Grumbine/runtime
 
 setenv x `date`
 echo start of loop at dtime $x
@@ -63,14 +64,15 @@ foreach yy ( 2011 )
           # -m cProfile -o pystats.$tag.$i to generate profiling 
           #   stats for later analysis. Optional
           python3  $EXDIR/setup_verf_ice.py $initial $tag ${FCST_BASE}/${initial}/6hrly/ >>& $expt.$initial.$tag.out
-          mv fcst_edge.$tag          $OUT
-          mv edge.fcst.*.*.$tag      $OUT
+          if ( -f fcst_edge.$tag ) then
+            mv fcst_edge.$tag          $OUT
+            mv edge.fcst.*.*.$tag      $OUT
+          endif
         end
         mv $expt.$initial.*.out $OUT
 
 # Per-forecast statistics, graphics
         time python3 $EXDIR/contingency_plots.py 35 $initial 0.15 
-        echo time python3 $EXDIR/contingency_plots.py 35 $initial 0.15 >> $OUT/to_run
         mv *${initial}.png                $OUT
         mv score*${initial}.csv           $OUT
       else
